requirements.txt
asgiref==3.3.4
astroid==2.5.2
asttokens==2.0.4
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.4
docutils==0.16
et-xmlfile==1.0.1
idna==2.10
isort==5.8.0
jedi==0.18.0
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.3
parso==0.8.1
Pillow<=8.0.0
pycodestyle==2.7.0
pylint==2.7.4
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
stripe==2.56.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wrapt==1.12.1
